move mock into tests.py
